pyproject.toml
[project]
name = "rag_juridico"
version = "0.1.0"
description = "Agente RAG constitucional utilizando LangChain, LangGraph, Pinecone e PostgreSQL"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langgraph",
    "psycopg2-binary",
    "openai",
    "pyyaml",
    "fastapi",
    "uvicorn",
    "tiktoken",
    "pandas",
    "numpy",
    "rich",
    "markdown>=3.9",
    "beautifulsoup4>=4.14.2",
    "pydantic>=2.12.3",
    "langchain-community>=0.4.1",
    "tqdm>=4.67.1",
    "python-dotenv>=1.2.1",
    "langchain-openai>=1.0.1",
    "pinecone>=6.0.0",
    "rank-bm25>=0.2.2",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.8",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "langchain-pinecone>=0.1.0",
    "fitz>=0.0.1.dev2",
    "docx>=0.2.4",
    "pymupdf>=1.26.5",
    "python-docx>=1.2.0",
    "pytesseract>=0.3.13",
    "pypdf2>=3.0.1",
    "pdfplumber>=0.11.7",
    "chroma>=0.2.0",
    "chromadb>=1.3.2",
    "pypdf>=6.1.3",
    "pillow>=12.0.0",
    "pdf2image>=1.17.0",
]

[project.urls]
Homepage = "https://www.perceptonIA.com.br/"

[tool.ruff]
line-length = 88
target-version = "py313"
fix = true
show-fixes = true
indent-width = 4
exclude = [".venv"]


[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "T201",   # Checks for print statements, 
  "COM812", # Checks for the absence of trailing commas.
  "INP001", # Checks for packages that are missing an __init__.py file.
  "D",      # All pydocstyle (D)
  "ANN401", # Checks that function arguments are annotated with a more specific type than Any.
  "ERA001", # Checks for commented-out Python code.
  "A004",   # Shadowing Python Builtin
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["rag_juridico"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.13"
include = ["src", "tests"]
exclude = [
  "**/venv",
  "**/.venv",
  "**/env",
  "**/.env",
  "**/node_modules",
  "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]

[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
